bt-311: + Changes to forgot-password form
