+++ frontend/src/bundles/auth/components/auth-forms/reset-password/styles.module.scss
@@ -1,0 +1,20 @@
+.header {
+    width: 100%;
+    margin-bottom: 25px;
+    font-size: 30px;
+}
+
+.input {
+    margin-bottom: 15px;
+}
+
+.btn {
+    font-size: 16px;
+}
+
+@media screen and (width >= 900px) {
+    .header {
+        font-size: 36px;
+        white-space: nowrap;
+    }
+}
