+++ frontend/src/bundles/auth/components/components.tsx
@@ -1,4 +1,4 @@
-export { ResetPassword } from '../pages/reset-password/reset-password.js';
+export { ResetPassword } from './auth-forms/reset-password/reset-password.js';
 export { SignInForm } from './auth-forms/sign-in-form/sign-in-form.js';
 export { SignUpForm } from './auth-forms/sign-up-form/sign-up-form.js';
 export { ProtectedRoute } from './protected-route/protected-route.js';
