+++ frontend/src/bundles/auth/pages/styles.scss
@@ -37,7 +37,7 @@
     font-family: var(--font-family);
 }
 
-.btnLogin {
+.btn {
     width: 100%;
     max-width: 420px;
     padding-top: 15px;
