+++ frontend/src/index.tsx
@@ -15,7 +15,6 @@ import {
 import { AppRoute } from '~/bundles/common/enums/enums.js';
 import { store } from '~/framework/store/store.js';
 
-import { ResetPassword } from './bundles/auth/pages/reset-password/reset-password.js';
 import { NotFoundPage } from './bundles/common/pages/not-found/not-found.js';
 import { theme } from './bundles/common/themes/theme.js';
 import { StepNavigation } from './bundles/talent-onboarding/components/components.js';
@@ -76,7 +75,7 @@ createRoot(document.querySelector('#root') as HTMLElement).render(
                             },
                             {
                                 path: AppRoute.RESET_PASSWORD,
-                                element: <ResetPassword />,
+                                element: <Auth />,
                             },
 
                             {
